+++ Src/Private/Diagram/Get-VbrServerConnection.ps1
@@ -13,7 +13,7 @@ function Get-VbrServerConnection {
         The TCP Port of the target Veeam Backup Server.
 
     .NOTES
-        Version:        0.6.38
+        Version:        0.8.24
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         GitHub:         rebelinux