+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :arrows_clockwise: Changed
 
 - Update module version to `v0.8.24`
-- Upgrade Veeam.Diagrammer module to `v0.6.36`
+- Upgrade Veeam.Diagrammer module to `v0.8.24`
 - Upgrade Diagrammer.Core module to `v0.2.36.1`
 - Update GitHub release workflow to use latest checkout action version
 - Update GitHub release workflow to use latest bluesky-post-action version
@@ -51,7 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :arrows_clockwise: Changed
 
 - Update module version to v0.8.23
-- Upgrade Veeam.Diagrammer module to v0.6.33
+- Upgrade Veeam.Diagrammer module to v0.8.24
 - Upgrade Diagrammer.Core module to v0.2.28
 
 ### :bug: Fixed
@@ -76,7 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :arrows_clockwise: Changed
 
 - Bump module to v0.8.22
-- Bump Veeam.Diagrammer module to v0.6.30
+- Bump Veeam.Diagrammer module to v0.8.24
 - Bump Diagrammer.Core module to v0.2.27
 - Update workflow to use Windows 2022 for publishing PowerShell module